Move provenance inside detail
