+++ context/app/static/js/components/Detail/provenance/ProvAnalysisDetailsLink/style.js
