+++ context/app/static/js/components/Detail/provenance/ProvGraph/style.js
