+++ context/app/static/js/components/Detail/provenance/ProvTable/style.js
