+++ context/app/static/js/components/Detail/provenance/ProvTabs/utils.spec.js
