+++ context/app/static/js/components/Detail/provenance/TabPanel/index.js
