+++ context/app/static/js/components/Detail/provenance/TabPanel/style.js
