+++ context/app/static/js/pages/Dataset/Dataset.jsx
@@ -3,7 +3,7 @@ import Typography from '@material-ui/core/Typography';
 
 import { LightBlueLink } from 'js/shared-styles/Links';
 import Files from 'js/components/files/Files';
-import ProvSection from 'js/components/provenance/ProvSection';
+import ProvSection from 'js/components/Detail/provenance/ProvSection';
 import Summary from 'js/components/Detail/Summary';
 import Attribution from 'js/components/Detail/Attribution';
 import Protocol from 'js/components/Detail/Protocol';
